# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CountsSummary": {
        "additionalProperties": false,
        "description": "Real per-run counts: events, spans, messages, LLM calls and tokens.",
        "properties": {
          "events": {
            "title": "Events",
            "type": "integer"
          },
          "llm_calls": {
            "title": "Llm Calls",
            "type": "integer"
          },
          "messages": {
            "title": "Messages",
            "type": "integer"
          },
          "spans": {
            "title": "Spans",
            "type": "integer"
          },
          "tokens": {
            "title": "Tokens",
            "type": "integer"
          }
        },
        "required": [
          "events",
          "spans",
          "messages",
          "llm_calls",
          "tokens"
        ],
        "title": "CountsSummary",
        "type": "object"
      },
      "DeterminismSummary": {
        "additionalProperties": false,
        "description": "A run's determinism health: canonical log hash and warning counts.",
        "properties": {
          "canonical_log_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Log Hash"
          },
          "nondeterminism_warnings": {
            "title": "Nondeterminism Warnings",
            "type": "integer"
          },
          "unwrapped_tools": {
            "title": "Unwrapped Tools",
            "type": "integer"
          }
        },
        "required": [
          "canonical_log_hash",
          "unwrapped_tools",
          "nondeterminism_warnings"
        ],
        "title": "DeterminismSummary",
        "type": "object"
      },
...
      "ExplorationResponse": {
        "additionalProperties": false,
        "description": "PRD \u00a726.1 / \u00a715.6. `coverage_statement` is I10's required, verbatim field.",
        "properties": {
          "capped": {
            "title": "Capped",
            "type": "boolean"
          },
          "coverage_statement": {
            "title": "Coverage Statement",
            "type": "string"
          },
          "k": {
            "title": "K",
            "type": "integer"
          },
          "new_findings": {
            "items": {
              "$ref": "#/components/schemas/FindingOut"
            },
            "title": "New Findings",
            "type": "array"
          },
          "reduced_away": {
            "title": "Reduced Away",
            "type": "integer"
          },
          "schedules_executed": {
            "title": "Schedules Executed",
            "type": "integer"
          },
          "unique_schedules": {
            "title": "Unique Schedules",
            "type": "integer"
          }
        },
        "required": [
          "k",
          "schedules_executed",
          "unique_schedules",
          "reduced_away",
          "new_findings",
          "capped",
          "coverage_statement"
        ],
        "title": "ExplorationResponse",
        "type": "object"
      },
...
      "GraphRef": {
        "additionalProperties": false,
        "description": "A run's graph identity \u2014 content hash plus the observed agent roster.",
        "properties": {
          "agents": {
            "items": {
              "type": "string"
            },
            "title": "Agents",
            "type": "array"
          },
          "hash": {
            "title": "Hash",
            "type": "string"
          }
        },
        "required": [
          "hash",
          "agents"
        ],
        "title": "GraphRef",
        "type": "object"
      },
      "GraphResponse": {
        "additionalProperties": false,
        "description": "`GET /api/runs/{id}/graph` response: nodes, edges and the critical path.",
        "properties": {
          "critical_path": {
            "items": {
              "type": "string"
            },
            "title": "Critical Path",
            "type": "array"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "title": "Nodes",
            "type": "array"
          }
        },
        "required": [
          "nodes",
          "edges",
          "critical_path"
        ],
        "title": "GraphResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunDetail": {
        "additionalProperties": false,
        "description": "`GET /api/runs/{id}` response (PRD \u00a726.1, `[SOURCE]`).\n\n`verdict` is `None` until the run's scorecard has been computed and persisted \u2014 no\nprompt in this build yet runs the full `timing -> overhead -> redundancy -> race ->\nbaseline -> resilience -> verdict` pipeline and calls `store.upsert_scorecard` (P17,\nnot started; see `docs/api.md` \"What this endpoint does not yet do\"). This is reported\nas `null`, not fabricated and not a 404 \u2014 PRD \u00a736 rule 2, \"partial results beat no\nresults\": the run's own metadata is always real, even when its verdict is not yet in.",
        "properties": {
          "baseline_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Run Id"
          },
          "comparability": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comparability"
          },
          "counts": {
            "$ref": "#/components/schemas/CountsSummary"
          },
          "determinism": {
            "$ref": "#/components/schemas/DeterminismSummary"
          },
          "graph": {
            "$ref": "#/components/schemas/GraphRef"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "scenario": {
            "$ref": "#/components/schemas/ScenarioRef"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "timing": {
            "$ref": "#/components/schemas/TimingSummary"
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerdictOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "run_id",
          "status",
          "mode",
          "seed",
          "scenario",
          "graph",
          "timing",
          "counts",
          "determinism",
          "verdict",
          "baseline_run_id",
          "comparability"
        ],
        "title": "RunDetail",
        "type": "object"
      },
...
      "ScenarioRef": {
        "additionalProperties": false,
        "description": "Which scenario a run used \u2014 id (if known) plus its resolved content hash.",
        "properties": {
          "hash": {
            "title": "Hash",
            "type": "string"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          }
        },
        "required": [
          "id",
          "hash"
        ],
        "title": "ScenarioRef",
        "type": "object"
      },
...
      "StateResponse": {
        "additionalProperties": false,
        "description": "`GET /api/runs/{id}/state` response.",
        "properties": {
          "at_seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "At Seq"
          },
          "at_virtual_ts": {
            "title": "At Virtual Ts",
            "type": "integer"
          },
          "keys": {
            "items": {
              "$ref": "#/components/schemas/StateKey"
            },
            "title": "Keys",
            "type": "array"
          }
        },
        "required": [
          "at_virtual_ts",
          "at_seq",
          "keys"
        ],
        "title": "StateResponse",
        "type": "object"
      },
      "TimingSummary": {
        "additionalProperties": false,
        "description": "Makespan figures for `GET /runs/{id}` \u2014 virtual, wall and critical-path ms.",
        "properties": {
          "critical_path_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Critical Path Ms"
          },
          "virtual_makespan_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Virtual Makespan Ms"
          },
          "wall_makespan_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wall Makespan Ms"
          }
        },
        "required": [
          "virtual_makespan_ms",
          "wall_makespan_ms",
          "critical_path_ms"
        ],
        "title": "TimingSummary",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WaterfallLane": {
        "additionalProperties": false,
        "description": "One agent's lane of spans in a `GET /api/runs/{id}/waterfall` response.",
        "properties": {
          "agent": {
            "title": "Agent",
            "type": "string"
          },
          "spans": {
            "items": {
              "$ref": "#/components/schemas/WaterfallSpan"
            },
            "title": "Spans",
            "type": "array"
          }
        },
        "required": [
          "agent",
          "spans"
        ],
        "title": "WaterfallLane",
        "type": "object"
      },
      "WaterfallResponse": {
        "additionalProperties": false,
        "description": "`GET /api/runs/{id}/waterfall` response: makespans plus every lane.",
        "properties": {
          "baseline_makespan_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Makespan Ms"
          },
          "lanes": {
            "items": {
              "$ref": "#/components/schemas/WaterfallLane"
            },
            "title": "Lanes",
            "type": "array"
          },
          "virtual_makespan_ms": {
            "title": "Virtual Makespan Ms",
            "type": "integer"
          }
        },
        "required": [
          "virtual_makespan_ms",
          "baseline_makespan_ms",
          "lanes"
        ],
        "title": "WaterfallResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Multi-agent coordination debugger, deterministic replay runtime and chaos harness \u2014 PRD \u00a726 API layer. Bounded search: absence of findings is not proof of absence.",
    "title": "AgentDX API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs/{run_id}": {
      "get": {
        "description": "Return a run's metadata, real counts and (when computed) its verdict.\n\nRaises:\n    RunNotFoundError: `E-RUN-404`.\n\n`timing.critical_path_ms` is reported as `null`: computing it means building the full\ntiming DAG (`analysis.timing.build_timing_dag` + `critical_path`), the same cost\n`GET .../graph` already pays where the PRD's own example payload needs the number \u2014 this\nendpoint does not additionally pay that cost on every call just to fill one field callers\nneeding it can get from `.../graph` directly. `determinism.unwrapped_tools` is `0` on\nevery run: no event or field anywhere in this codebase currently records it (grepped, not\nassumed) \u2014 a real, declared gap, not a fabricated zero passed off as a real count.",
        "operationId": "get_run_detail_api_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run metadata + verdict (PRD \u00a726.1)",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/runs/{run_id}/exploration": {
      "get": {
        "description": "Return a run's exploration `Report` (PRD \u00a715), including the required `coverage_statement`.\n\nRaises:\n    RunNotFoundError: `E-RUN-404`.\n    NotYetAvailableError: `409` \u2014 always, in this build. `explore.report.Report` is real\n        (P13) but `api/` may not import `agentdx.explore` (`.importlinter`\n        `explore-below-transport-layer`: `explore/` must not import `api/`, and the\n        converse contract, `api-never-imports-runtime`, was written before `explore/`\n        existed and does not list it as permitted either \u2014 confirmed against the actual\n        `.importlinter`, not assumed from the CONTEXT.md \u00a74 table's prose). No prompt\n        persists an exploration `Report` into the store (no table exists for one \u2014 see\n        `docs/api.md`), so there is nothing this route can read regardless. Surfaced\n        plainly rather than fabricated (PRD \u00a736 rule 1).",
        "operationId": "get_exploration_api_runs__run_id__exploration_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplorationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bounded schedule exploration results (PRD \u00a726.1, \u00a715.6)",
        "tags": [
          "analysis"
        ]
      }
    },
...
    "/api/runs/{run_id}/graph": {
      "get": {
        "description": "Return nodes, edges and the critical path, computed live from the sealed event log.\n\nRaises:\n    RunNotFoundError: `E-RUN-404`.\n    TimingAnalysisError / OverheadAnalysisError: `409` \u2014 the log is not yet in a shape\n        this analysis can run over (PRD \u00a736 `E-ANLZ-*`; most commonly, the run has not\n        been sealed and still has an open span).\n\n`at_virtual_ts`, when given, restricts the *reported* nodes/edges to ones whose own\nactivity had started by that instant; the critical path and `cp_share`/`cp_ms` figures\nare still computed over the *full* sealed run and only filtered to the surviving node\nids \u2014 this is a simplification of PRD \u00a726.1's \"graph state as of that instant\", declared\nin `docs/api.md`, not a full historical re-derivation of what the critical path *would\nhave been* had the run stopped at that instant.",
        "operationId": "get_graph_api_runs__run_id__graph_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "description": "Restrict to nodes that had started by this virtual ms (PRD \u00a726.1).",
            "in": "query",
            "name": "at_virtual_ts",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Restrict to nodes that had started by this virtual ms (PRD \u00a726.1).",
              "title": "At Virtual Ts"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Node/edge graph (PRD \u00a726.1)",
        "tags": [
          "analysis"
        ]
      }
    },
...
    "/api/runs/{run_id}/state": {
      "get": {
        "description": "Return every state key's last-known value hash as of `at_virtual_ts` (PRD \u00a720.4).\n\nRaises:\n    RunNotFoundError: `E-RUN-404`.",
        "operationId": "get_state_at_api_runs__run_id__state_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "description": "Reconstruct state as of this virtual ms.",
            "in": "query",
            "name": "at_virtual_ts",
            "required": true,
            "schema": {
              "description": "Reconstruct state as of this virtual ms.",
              "title": "At Virtual Ts",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reconstructed state (PRD \u00a726.1, \u00a720.4)",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/runs/{run_id}/waterfall": {
      "get": {
        "description": "Return per-agent lanes of spans with virtual start/end and critical-path flags.\n\nRaises:\n    RunNotFoundError: `E-RUN-404`.\n    TimingAnalysisError: `409` \u2014 see `get_graph`.\n\n`bucket` (PRD \u00a716.2.2's six-way overhead classification) is reported as `null` on every\nspan: `analysis.overhead` computes and aggregates bucket *totals*\n(`OverheadDecomposition.bucket_ms`) but exposes no public per-node classification to\nproject back onto an individual span \u2014 only its private `_classify_node` does, and this\nlayer does not reach into another module's private surface. This is a real, declared\ngap (`docs/api.md` \"What GET .../waterfall does not yet do\"), not a silent omission.",
        "operationId": "get_waterfall_api_runs__run_id__waterfall_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "description": "Window start, virtual ms.",
            "in": "query",
            "name": "from_ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Window start, virtual ms.",
              "title": "From Ms"
            }
          },
          {
            "description": "Window end, virtual ms.",
            "in": "query",
            "name": "to_ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Window end, virtual ms.",
              "title": "To Ms"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaterfallResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Span waterfall (PRD \u00a726.1)",
        "tags": [
          "analysis"
        ]
      }
    },
...
  }
}